Fix windows batch to use correct directory
The python script is always referenced relative to the batch file (so
it's always found), but the data files are processed in their current
working directory. This allows you to drag a .png file onto the
convert.bat file from anywhere in Windows, and it will be converted into
a .bin file in place.