Update fonts to use new smoothing options
Also remove the px variants, as the kerning on them is just awful. Note
that I left the .bin files around on purpose so the build would not
break; someone who can test needs to switch the font files over in-game,
and add in the xo and yo offset logic.