Put everything in namespaces to avoid clashes with other libraries.
When trying to build the browser from scratch with the latest ctrulib, I
found that there was a conflicting declaration of a type that was in the
browser before that now also exists in ctrulib. To prevent this from
happening in the future, I went ahead and put everything in namespaces.